rename test project to improve clarity of intent
