+++ test/Spring/Spring.Data.NHibernate.NestedTxSuspension.Integration.Tests/App.config
@@ -15,8 +15,8 @@
   <spring>
     <context>
       <resource uri="config://spring/objects"/>
-      <resource uri="assembly://Spring.Data.NHibernate.TxPromotion.Integration.Tests/Spring.Data.NHibernate.TxPromotion.Integration.Tests/Spring.Configuration.xml"/>
-      <resource uri="assembly://Spring.Data.NHibernate.TxPromotion.Integration.Tests/Spring.Data.NHibernate.TxPromotion.Integration.Tests/Spring.NHibernate.xml"/>
+      <resource uri="assembly://Spring.Data.NHibernate.NestedTxSuspension.Integration.Tests/Spring.Data.NHibernate.NestedTxSuspension.Integration.Tests/Spring.Configuration.xml"/>
+      <resource uri="assembly://Spring.Data.NHibernate.NestedTxSuspension.Integration.Tests/Spring.Data.NHibernate.NestedTxSuspension.Integration.Tests/Spring.NHibernate.xml"/>
     </context>
     <parsers>
       <parser type="Spring.Data.Config.DatabaseNamespaceParser, Spring.Data"/>
