+++ test/Spring/Spring.Data.NHibernate.NestedTxSuspension.Integration.Tests/Mappings.hbm.xml
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
-<hibernate-mapping xmlns="urn:nhibernate-mapping-2.2" default-lazy="true" assembly="Spring.Data.NHibernate.TxPromotion.Integration.Tests" namespace="Spring.Data.NHibernate.TxPromotion.Integration.Tests">
+<hibernate-mapping xmlns="urn:nhibernate-mapping-2.2" default-lazy="true" assembly="Spring.Data.NHibernate.NestedTxSuspension.Integration.Tests" namespace="Spring.Data.NHibernate.NestedTxSuspension.Integration.Tests">
 
   <class name="Container" proxy="IContainer">
     <id name="Id">
